+++ packages/core/src/Models.ts
@@ -1,0 +1,12 @@
+export interface WithVisibilityOption {
+  isVisible?: boolean;
+}
+export interface WithActiveOption {
+  active?: boolean;
+}
+export interface OnChangeCustomEvent {
+  value: number;
+}
+export interface WithOnChangeEvent<T> {
+  onChange: (event: T) => void;
+}
