+++ packages/core/src/index.ts
@@ -9,4 +9,8 @@ export {
   WithClickIdProps,
 } from './withClickId.hoc';
 export { default as withClassDefault } from './withClassDefault.hoc';
-export { default as withClassModifier, WithClassModifierOptions } from './withClassModifier.hoc';
+export {
+  default as withClassModifier,
+  WithClassModifierOptions,
+} from './withClassModifier.hoc';
+export * from './Models';
