+++ packages/table/src/Pager/Previous.tsx
@@ -4,7 +4,7 @@ import {
   WithVisibilityOption,
   WithOnChangeEvent,
   OnChangeCustomEvent,
-} from './Pager';
+} from '@axa-fr/react-toolkit-core';
 
 export type PreviousProps = WithActiveOption &
   WithVisibilityOption &
