+++ storybook/storybook/src/packages/table/src/Pager/Pager.stories.js
@@ -3,7 +3,7 @@ import { action } from '@storybook/addon-actions';
 import { text, number, select } from '@storybook/addon-knobs';
 
 import addToStorie from '@axa-fr/storybook-addons';
-import { Pager } from '@axa-fr/react-toolkit-table';
+import { Pager, Modes } from '@axa-fr/react-toolkit-table';
 
 const readme = require('@axa-fr/react-toolkit-table/dist/Pager/README.md');
 
@@ -17,7 +17,7 @@ stories.push({
       onChange={action('onChange')}
       classModifier={text('classModifier', '')}
       className={text('className', '')}
-      mode={select('mode', Pager.Modes, Pager.Modes.default)}
+      mode={select('mode', Modes, Modes.default)}
     />
   ),
 });
